# Conflicts:
#	composer.lock
#	package-lock.json

+++ src/Console/Command/PurgeCommand.php
@@ -24,24 +24,24 @@ use Symfony\Component\Finder\SplFileInfo;
 
 class PurgeCommand extends BaseCommand
 {
-    protected function configure()
+    protected function configure(): void
     {
         $this->setName('purge')
             ->setDescription('Purge packages')
             ->setDefinition([
                 new InputArgument('file', InputArgument::OPTIONAL, 'Json file to use', './satis.json'),
                 new InputArgument('output-dir', InputArgument::OPTIONAL, 'Location where to output built files', null),
+                new InputArgument('dry-run', InputArgument::OPTIONAL, 'Dry run, allows to inspect what might be deleted', null),
             ])
             ->setHelp(
-<<<'EOT'
-The <info>purge</info> command deletes useless archive files, depending
-on given json file (satis.json is used by default) and the
-newest json file in the include directory of the given output-dir.
+                <<<'EOT'
+                The <info>purge</info> command deletes useless archive files, depending
+                on given json file (satis.json is used by default) and the
+                newest json file in the include directory of the given output-dir.
 
-In your satis.json (or other name you give), you must define
-"archive" argument.
-
-EOT
+                In your satis.json (or other name you give), you must define
+                "archive" argument. You also need to define "homepage" argument or "SATIS_HOMEPAGE" environment variable if you don't use archive "prefix-url" argument.
+                EOT
             );
     }
 
@@ -70,12 +70,17 @@ EOT
             throw new \InvalidArgumentException('The output dir must be specified as second argument or be configured inside ' . $input->getArgument('file'));
         }
 
+        $dryRun = (bool) $input->getArgument('dry-run');
+        if ($dryRun) {
+            $output->writeln('<notice>Dry run enabled, no actual changes will be done.</notice>');
+        }
+
         $packageSelection = new PackageSelection($output, $outputDir, $config, false);
         $packages = $packageSelection->load();
 
         $prefix = sprintf(
             '%s/%s/',
-            $config['archive']['prefix-url'] ?? $config['homepage'],
+            $config['archive']['prefix-url'] ?? getenv('SATIS_HOMEPAGE') ?: $config['homepage'],
             $config['archive']['directory']
         );
 
@@ -121,7 +126,9 @@ EOT
         }
 
         foreach ($unreferenced as $file) {
-            unlink($file->getPathname());
+            if (!$dryRun) {
+                unlink($file->getPathname());
+            }
 
             $output->writeln(sprintf(
                 '<info>Removed archive</info>: <comment>%s</comment>',
@@ -129,7 +136,9 @@ EOT
             ));
         }
 
-        $this->removeEmptyDirectories($output, $distDirectory);
+        if (!$dryRun) {
+            $this->removeEmptyDirectories($output, $distDirectory);
+        }
 
         $output->writeln('<info>Done.</info>');
